feat(ui) Display consistent search results in embedded searches
